Move setupWorld to separate file
