+++ src/Constants/enemies.js
@@ -1,0 +1,5 @@
+import club from './club';
+
+const enemies = [club.luzern, club.basel, club.stGallen];
+
+export { enemies as default };
