+++ src/Game.jsx
@@ -1,22 +1,15 @@
-import Matter from 'matter-js';
 import React, { useState, useRef } from 'react';
 import { GameEngine } from 'react-game-engine';
-import Background from './Entities/Background';
-import Ball from './Entities/Ball';
-import Constants, { GetAbsolutHeightPosition, GetAbsolutWidthPosition } from './Constants';
-import Floor from './Entities/Floor';
-import Physics from './Physics/Physics';
-import YbPlayer from './Entities/YbPlayer';
-import addEnemies from './Entities/enemies';
-import club from './Constants/club';
-import skinColor from './Constants/skinColor';
-import levelUpSound from './Assets/Sounds/levelup.mp3';
-import gameOverSound from './Assets/Sounds/gameOver.mp3';
-import headerSound from './Assets/Sounds/header.mp3';
-import pokal from './Assets/Images/Pokal.jpg';
+import Constants from './Constants';
 import GameOver from './GameOver';
 import LevelUp from './LevelUp';
+import Physics from './Physics/Physics';
 import Welcome from './Welcome';
+import enemies from './Constants/enemies';
+import gameOverSound from './Assets/Sounds/gameOver.mp3';
+import levelUpSound from './Assets/Sounds/levelup.mp3';
+import pokal from './Assets/Images/Pokal.jpg';
+import setupWorld from './setupWorld';
 
 const styles = {
   container: {
@@ -96,12 +89,10 @@ const styles = {
   },
 };
 
-const clubOrder = [club.luzern, club.basel, club.stGallen];
-
-const SimpleGame = () => {
+const Game = () => {
   const gameEngine = useRef(null);
 
-  const [clubIndex, setClubIndex] = useState(0);
+  const [enemyIndex, setEnemyIndex] = useState(0);
   const [isGoalReached, setGoalReached] = useState(false);
   const [isEndReached, setEndReached] = useState(false);
   const [isWelcomeScreenVisible, showWelcomeScreen] = useState(true);
@@ -109,131 +100,39 @@ const SimpleGame = () => {
   const [score, setScore] = useState(0);
   const [numberOfTrophies, setNumberOfTrophies] = useState(0);
   const [year, setYear] = useState(2018);
+  const [entities] = useState(setupWorld(gameEngine, isGoalReached, enemies[enemyIndex]));
 
   const levelUpAudio = new Audio(levelUpSound);
   const gameOverAudio = new Audio(gameOverSound);
-  const headerAudio = new Audio(headerSound);
-
-  const bodiesAreColliding = (pair, nameA, nameB) => {
-    const isAColidingWithB = (pair.bodyB.label === nameA && pair.bodyA.label === nameB);
-    const isBColidingWithA = (pair.bodyB.label === nameB && pair.bodyA.label === nameA);
-    return isAColidingWithB || isBColidingWithA;
-  };
-
-  const setupWorld = () => {
-    const engine = Matter.Engine.create();
-    const { world } = engine;
-    world.gravity.y = Constants.GRAVITY;
-    const ball = Matter.Bodies.circle(Constants.MAX_WIDTH / 2, 0, GetAbsolutHeightPosition(3));
-    const stadium = Matter.Bodies.rectangle(0, 0, 0, 0);
-    const city = Matter.Bodies.rectangle(0, 0, 0, 0);
-
-    ball.label = 'ball';
-    ball.isNotFixed = false;
-
-    const floor1 = Matter.Bodies.rectangle(
-      Constants.MAX_WIDTH / 2,
-      Constants.MAX_HEIGHT - GetAbsolutHeightPosition(10),
-      Constants.MAX_WIDTH + 4,
-      GetAbsolutHeightPosition(10),
-      { isStatic: true },
-    );
-    floor1.label = 'floor';
-    const player1 = Matter.Bodies.rectangle(
-      Constants.MAX_WIDTH / 2,
-      Constants.MAX_HEIGHT - GetAbsolutHeightPosition(20),
-      GetAbsolutWidthPosition(10),
-      GetAbsolutHeightPosition(20),
-      {
-        inertia: Infinity,
-      },
-    );
-    player1.label = 'player1';
-    player1.mass = 1;
-    player1.inverseMass = 1;
-
-    Matter.World.add(world, [ball, floor1, player1]);
-    Matter.Events.on(engine, 'collisionStart', (event) => {
-      if (event.pairs.filter((element) => bodiesAreColliding(element, 'player1', 'ball')).length !== 0) {
-        if (isGoalReached) {
-          ball.isNotFixed = true;
-          Matter.Body.setVelocity(ball, {
-            x: GetAbsolutWidthPosition(1),
-            y: ball.velocity.y,
-          });
-        } else {
-          Matter.Body.setVelocity(ball, {
-            x: ball.velocity.x,
-            y: -GetAbsolutWidthPosition(1),
-          });
-        }
-        headerAudio.play();
-        gameEngine.current.dispatch({ type: 'score' });
-      } else if (event.pairs.filter((element) => bodiesAreColliding(element, 'floor', 'ball')).length !== 0 && !isGoalReached) {
-      // The ball collided with the floor.
-        gameEngine.current.dispatch({ type: 'game-over' });
-      }
-    });
-    const enemies = addEnemies(engine, clubOrder[clubIndex]);
-
-    return {
-      ...enemies,
-      physics: { engine, world },
-      ball: { body: ball, renderer: Ball },
-      floor1: { body: floor1, renderer: Floor },
-      player1: {
-        body: player1,
-        stadium,
-        city,
-        renderer: YbPlayer,
-        skinColor: skinColor.latin,
-        club: club.yb,
-      },
-
-      bg: {
-        stadium,
-        city,
-        endReached: () => {
-          gameEngine.current.dispatch({ type: 'end-reached' });
-        },
-        goalReached: () => {
-          gameEngine.current.dispatch({ type: 'goal-reached' });
-        },
-        renderer: Background,
-      },
-    };
-  };
-
-  const [entities] = useState(setupWorld);
 
   const handleContinue = () => {
     runGame(true);
     setScore(score + 100);
     setYear(year + 1);
     setNumberOfTrophies(numberOfTrophies + 1);
-    setClubIndex(clubIndex + 1);
+    setEnemyIndex(enemyIndex + 1);
 
-    if (clubIndex >= clubOrder.length) {
-      setClubIndex(0);
+    if (enemyIndex >= enemies.length) {
+      setEnemyIndex(0);
     }
 
-    gameEngine.current.swap(setupWorld());
+    gameEngine.current.swap(setupWorld(gameEngine, isGoalReached, enemies[enemyIndex]));
   };
 
   const handleReset = () => {
-    gameEngine.current.swap(setupWorld());
+    gameEngine.current.swap(setupWorld(gameEngine, isGoalReached, enemies[enemyIndex]));
 
     runGame(true);
     setScore(0);
     setYear(2018);
     setNumberOfTrophies(0);
-    setClubIndex(0);
+    setEnemyIndex(0);
   };
 
   const handleEvent = (e) => {
     if (e.type === 'game-over') {
       // Alert.alert("Game Over");
-      setClubIndex(0);
+      setEnemyIndex(0);
       gameOverAudio.play();
       runGame(false);
     } else if (e.type === 'score') {
@@ -302,4 +201,4 @@ const SimpleGame = () => {
   );
 };
 
-export default SimpleGame;
+export default Game;
